jstl 적용 전

+++ src/main/webapp/jsp/usr/article/list.jsp
@@ -1,19 +1,27 @@
 <%@ page language="java" contentType="text/html; charset=UTF-8" pageEncoding="UTF-8"%>
 <%@ page import="java.util.List" %>
 <%@ page import="com.ll.exam.article.Dto.ArticleDto" %>
-
+<%@ taglib uri="http://java.sun.com/jsp/jstl/core" prefix="c"%>
 <%
     List<ArticleDto> articleDtoList = (List<ArticleDto>)request.getAttribute("articleDtoList");
 %>
 
-<h1>게시물 리스트</h1>
+<%@ include file="../common/head.jspf"%>
 
-<ul>
-    <% for ( ArticleDto article : articleDtoList ) { %>
-    <li>
-        <a href="/usr/article/detail/free/<%=article.getId()%>"><%=article.getId()%>. <%=article.getTitle()%> </a>
-        <a href="/usr/article/modify/free/<%=article.getId()%>">수정</a>
-        <a href="/usr/article/delete/free/<%=article.getId()%>">삭제</a>
-    </li>
-    <% } %>
-</ul>
+<section>
+    <div class="container px-3 mx-auto">
+        <h1 class="font-bold text-lg">게시물 리스트</h1>
+
+        <ul class="mt-5">
+            <% for ( ArticleDto article : articleDtoList ) { %>
+            <li class="flex">
+                <a class="w-[40px] hover:underline hover:text-[red]" href="/usr/article/detail/free/<%=article.getId()%>"><%=article.getId()%></a>
+                <a class="flex-grow hover:underline hover:text-[red]" href="/usr/article/detail/free/<%=article.getId()%>"><%=article.getTitle()%></a>
+                <a onclick="if ( !confirm('정말로 삭제하시겠습니까?') ) return false;" class="hover:underline hover:text-[red] mr-2" href="/usr/article/delete/free/<%=article.getId()%>">삭제</a>
+                <a class="hover:underline hover:text-[red]" href="/usr/article/modify/free/<%=article.getId()%>">수정</a>
+            </li>
+            <% } %>
+        </ul>
+    </div>
+</section>
+<%@ include file="../common/footer.jspf"%>
